requirements.txt
mlxtend==0.18.0
sklearn_genetic==0.4.1
geneticalgorithm==1.0.2
scipy==1.6.3
matplotlib==3.4.1
numpy==1.18.4
pandas==1.2.4
scikit_learn==0.24.2
